feat: add a basic Kubernetes manifest
Add a basic kustomize Kubernetes manifest derived from the template
output of the original power-control chart:
https://github.com/Cray-HPE/hms-power-control-charts/blob/main/charts/v2.2/cray-power-control/Chart.yaml

The source chart was modified to remove the original etcd subchart, and
"ochami-power-control-" replaces the original "cray-power-control-" in
the output resources.

The Helm release name was "riptide", as a recognizable string unlikely
to appear in the template output itself.

Signed-off-by: Travis Raines <571832+rainest@users.noreply.github.com>