wip: condense ConfigMap into Deployment
Remove the single-value ConfigMap and set PCS_CA_URI to its value
directly in the Deployment. Didn't see any obvious reason to keep this
separate; I don't think power-control can live-reload it, and in any
case it's a URL and can update upstream.

The existing CSM environments haven't changed it in 666 days (very metal
of them) and it still uses the default empty value. It doesn't appear
that important.

Signed-off-by: Travis Raines <571832+rainest@users.noreply.github.com>